Add new options to filter and limit loads (#19)
Two new runtime tunable options are added:
1. The number of loads to pass to the client
2. The minimum value for call minutes to pass to the client

These settings are meant to supersede the setting in Burble that
controls the number of loads displayed by manifest. The server is still
limited by this setting, so Burble should be set to one greater than the
number configured here. Loads that have call times less than the minimum
call minutes is also meant to supersede the related Burble setting.

The motivation for these changes is to help with the turning jumpers
feature. Since we want turning jumpers to be marked on the load they're
turning on, we must keep information about their previous load around.
The server can only keep a snapshot of data from Burble, so if/when the
previous load disappears from the Burble data source, we can no longer
determine whether the jumper is turning or not. So, keep the previous
load around longer than normal, but let it fall off the display sooner.

We can't keep information about prior loads that have disappeared from
the Burble data source hanging around, because when they disappear we
have no way of knowing _why_ they disappeared. We can assume, but then
we'll likely get that assumption wrong some of the time, and
@kdrivas1989 will be blowing up my phone with texts about things not
working right 😉