Add support for client session management (#24)
This adds support for client session management. It uses sqlite3 for a
backend to store information about users and sessions. It uses Sign in
with Apple for user authentication. The ability to easily add support
for other database backends or authentication providers is baked into
the implementation. This extends the gRPC protocol to add support for
signing in, signing out, and verifying sessions. Other features
dependent on authorization will come later.

With respect to Sign in with Apple, there is some code left stubbed out
here. Apple supports server-to-server notifications to advise
applications when users do certain things via appleid.apple.com. The
support to handle this has been added on a development branch in the
orangematt/siwa codebase, but testing it so far has proved troublesome
because Apple's servers won't connect to Golang's TLS implementation for
some reason and I haven't had the bandwidth to properly diagnose and
debug the problem. This is a pretty low-priority feature anyway.